Warning free on GCC 4.3.2

+++ parse.h
@@ -6,7 +6,7 @@
 
 #define NO_DPN          ((D_ParseNode*)0x1)
 #define DPN_TO_PN(_dpn) \
-  ((PNode *)(((char*)dpn)-(int)(&((PNode*)0)->parse_node)))
+  ((PNode *)(((char*)dpn)-(intptr_t)(&((PNode*)0)->parse_node)))
 #define is_epsilon_PNode(_pn) \
 ((_pn)->parse_node.start_loc.s == (_pn)->parse_node.end)
